+++ src/python_package/__main__.py
@@ -5,15 +5,26 @@
 
 
 import sys
+
 from . import PACKAGE_NAME, __version__
-
-
-def package_info():
-    return f'{PACKAGE_NAME} {__version__}'
+from .cli import CommandLineInterface
 
 
 def main():
-    print(package_info())
+    with CommandLineInterface(
+        label=PACKAGE_NAME,
+        version=__version__,
+    ) as app:
+        try:
+            app.run()
+
+        except AssertionError as e:
+            print('AssertionError > %s' % e.args[0])
+            app.exit_code = 1
+
+            if app.debug is True:
+                import traceback
+                traceback.print_exc()
 
 
 if __name__ == '__main__':